+++ api-builder-plugin-fn-elasticsearch/src/actions/indexTemplate.js
@@ -26,32 +26,19 @@ async function getTemplate(params, options) {
 	if (!params.name) {
 		throw new Error('Missing required parameter: name');
 	}
+	try {
+		var client = new ElasticsearchClient(elasticSearchConfig).client;
+		var indexTemplate = await client.indices.getTemplate( params, { ignore: [404], maxRetries: 3 });
 
-	var client = new ElasticsearchClient(elasticSearchConfig).client;
-	var indexTemplate = await executeRequest(params);
+		if(Object.keys(indexTemplate.body).length === 0 && indexTemplate.body.constructor === Object) {
+			return options.setOutput('notFound', `No index template found with name [${params.name}]`);
+		}
+		// Return the template config itself - Not the surrounding object based on the template name
+		return indexTemplate.body[params.name];
 
-	if(Object.keys(indexTemplate.body).length === 0 && indexTemplate.body.constructor === Object) {
-		return options.setOutput('notFound', `No index template found with name [${params.name}]`);
-	}
-	// Return the template config itself - Not the surrounding object based on the template name
-	return indexTemplate.body[params.name];
-
-	function executeRequest(params) {
-		return new Promise((resolve, reject) => {
-			client.indices.getTemplate( params, { ignore: [404], maxRetries: 3 }, (err, result) => {
-				if(err) {
-					if(!err.body) {
-						options.logger.error(`Error returned from Elastic-Search: ${JSON.stringify(err)}`);
-					}
-					reject(err.body.error.root_cause[0].reason);
-				} else if(result.error) {
-					reject(result.error);
-				} else {
-					resolve(result);
-				}
-			});
-	
-		})
+	} catch (e) {
+		if(e instanceof Error) throw e;
+		throw new Error(JSON.stringify(e));
 	}
 }
 
@@ -69,27 +56,46 @@ async function putTemplate(params, options) {
 		throw new Error('Missing required parameter: body');
 	}
 
-	var client = new ElasticsearchClient(elasticSearchConfig).client;
-	var result = await executeRequest();
+	var updateWhenChanged = false;
+	if(params.updateWhenChanged != undefined) {
+		updateWhenChanged = params.updateWhenChanged;
+		delete params.updateWhenChanged;
+	}
+	var recreate = params.recreate;
+	delete params.recreate;
 
-	return result;
-
-	function executeRequest() {
-		return new Promise((resolve, reject) => {
-			client.indices.putTemplate( params, { ignore: [404], maxRetries: 3 }, (err, result) => {
-				if(err) {
-					if(!err.body) {
-						options.logger.error(`Error returned from Elastic-Search: ${JSON.stringify(err)}`);
+	try {
+		var client = new ElasticsearchClient(elasticSearchConfig).client;
+		var actualTemplate;
+		// Use the index-template version to decide if the index should be updated or not
+		if(updateWhenChanged) {
+			// Get the current index-template based on the name
+			const response = await client.indices.getTemplate({name: params.name}, { ignore: [404], maxRetries: 3 });
+			actualTemplate = response.body[params.name];
+			if(actualTemplate == undefined) {
+				options.logger.info(`No index template found with name: ${params.name} creating new.`);
+			} else {
+				if(actualTemplate.version >= params.body.version) {
+					if(recreate) {
+						options.logger.info(`Desired index template version: ${params.body.version} is less or equal to actual version: ${actualTemplate.version}. Recreating index template anyway, as recreate is set.`);
+					} else {
+						return options.setOutput('noUpdate', `No update required as actual index template version: ${actualTemplate.version} is less or equal to desired version: ${params.body.version}`);
 					}
-					reject(err.body.error.root_cause[0].reason);
-				} else if(result.error) {
-					reject(result.error);
-				} else {
-					resolve(result);
 				}
-			});
-	
-		})
+			}
+		}
+		if(actualTemplate != undefined) {
+			// Take over existing aliases from the actual template
+			params.body.aliases = {...actualTemplate.aliases};
+			// Take over existing ILM Policies
+			params.body.settings.index.lifecycle = {...actualTemplate.settings.index.lifecycle};
+		}
+		var result = await client.indices.putTemplate( params, { ignore: [404], maxRetries: 3 });
+
+		return result;
+	} catch (e) {
+		if(e instanceof Error) throw e;
+		throw new Error(JSON.stringify(e));
 	}
 }